Changes to the Sherlock Exceptions.
