+++ sherlock/exception/__init__.py
@@ -1,0 +1,2 @@
+from sherlock.exception.slexception import SLException
+from sherlock.exception.slunsupportedtypeexception import SLUnsupportedTypeException
